+++ app/Models/RequestLog.php
@@ -24,22 +24,52 @@ class RequestLog extends Model
         return count($this->select('id')->get());
     }
 
-    public function paginatedCollection(int $offset = 0, int $limit = 10)
+    /**
+     * @param int $offset
+     * @param int $limit
+     * @param array $collection_parameters
+     *
+     * @return array
+     */
+    public function paginatedCollection(
+        int $offset = 0,
+        int $limit = 10,
+        array $collection_parameters = []
+    )
     {
-        return $this->orderByDesc('created_at')->offset($offset)->limit($limit)->get();
+        $collection = $this->orderByDesc('created_at');
+
+        if (array_key_exists('source', $collection_parameters) === true) {
+            $collection->where('source', '=', $collection_parameters['source']);
+        }
+
+        return $collection->offset($offset)->
+            limit($limit)->
+            get()->
+            toArray();
     }
 
-    public function monthlyRequests()
+    /**
+     * @param array $collection_parameters
+     *
+     * @return array
+     */
+    public function monthlyRequests(array $collection_parameters = [])
     {
         $collection = $this->orderBy(DB::raw("DATE_FORMAT(`request_log`.`created_at`, '%Y-%m')"))
-            ->groupBy(DB::raw("DATE_FORMAT(`request_log`.`created_at`, '%Y-%m')"))
-            ->select(
-                DB::raw("DATE_FORMAT(`request_log`.`created_at`, '%Y-%m')"),
-                DB::raw("COUNT(`request_log`.`id`) AS `requests`"),
-                DB::raw("ANY_VALUE(DATE_FORMAT(`request_log`.`created_at`, '%Y')) AS `year`"),
-                DB::raw("ANY_VALUE(DATE_FORMAT(`request_log`.`created_at`, '%M')) AS `month`")
-            );
+            ->groupBy(DB::raw("DATE_FORMAT(`request_log`.`created_at`, '%Y-%m')"));
 
-        return $collection->get();
+        if (array_key_exists('source', $collection_parameters) === true) {
+            $collection->where("source", '=', $collection_parameters['source']);
+        }
+
+        $collection->select(
+            DB::raw("DATE_FORMAT(`request_log`.`created_at`, '%Y-%m')"),
+            DB::raw("COUNT(`request_log`.`id`) AS `requests`"),
+            DB::raw("ANY_VALUE(DATE_FORMAT(`request_log`.`created_at`, '%Y')) AS `year`"),
+            DB::raw("ANY_VALUE(DATE_FORMAT(`request_log`.`created_at`, '%M')) AS `month`")
+        );
+
+        return $collection->get()->toArray();
     }
 }