+++ app/Models/Transformers/RequestLog.php
@@ -6,27 +6,31 @@ namespace App\Models\Transformers;
 /**
  * Transform the data returns from Eloquent into the format we want for the API
  *
+ * This is an updated version of the transformers, the other transformers need to
+ * be updated to operate on an array rather than collections
+ *
  * @author Dean Blackborough <dean@g3d-development.com>
  * @copyright Dean Blackborough 2018-2019
  * @license https://github.com/costs-to-expect/api/blob/master/LICENSE
  */
 class RequestLog extends Transformer
 {
-    protected $log;
+    protected $data_to_transform;
 
-    public function __construct(\App\Models\RequestLog $log)
+    public function __construct(array $data_to_transform)
     {
         parent::__construct();
 
-        $this->log = $log;
+        $this->data_to_transform = $data_to_transform;
     }
 
     public function toArray(): array
     {
         return [
-            'method' => $this->log->method,
-            'request_uri' => $this->log->request,
-            'created' => $this->log->created_at->toDateTimeString()
+            'method' => $this->data_to_transform['method'],
+            'source' => $this->data_to_transform['source'],
+            'request_uri' => $this->data_to_transform['request'],
+            'created' => $this->data_to_transform['created_at']
         ];
     }
 }