+++ resources/lang/en/request/parameters.php
@@ -1,0 +1,8 @@
+<?php
+
+declare(strict_types=1);
+
+return [
+    'title-source' => 'Source',
+    'description-source' => 'Show results for the requested source, api, legacy or website'
+];